repeatable read isolation level
the same as read committed but with the following anomaly not allowed:
P2 (“Non-repeatable read”): SQL-transaction T1 reads a row. SQL-transaction T2 then modifies or deletes that row and performs a COMMIT.
If T1 then attempts to reread the row, it may receive the modified value or discover that the row has been deleted. Thus breaking the consistency guarantee from ACID properties.